gcp/cloud_functions/requirements.txt
import functions_framework
firebase_functions
firebase-admin
requests
google-cloud-storage
tensorflow-cpu==2.12.1
Pillow==6.0.0